requirements.txt
faiss-cpu
numpy
requests
sentence-transformers
fastapi
uvicorn
